Dialogs retain their presenters too
Dialogs retain their presenters too.
FeedbackDialog now gets a more detailed context that activities can provide via baseActivity.provideDetailedFeedbackContext. MainActivity returns info about the currently visible fragment.

+++ app/src/main/java/org/schulcloud/mobile/ui/base/BaseFragment.java
@@ -77,7 +77,7 @@ public abstract class BaseFragment<V extends MvpView, P extends BasePresenter<V>
     @Override
     public void onSaveInstanceState(Bundle outState) {
         super.onSaveInstanceState(outState);
-        outState.putLong(KEY_ACTIVITY_ID, mActivityId);
+        outState.putInt(KEY_ACTIVITY_ID, mActivityId);
     }
 
     @NonNull
